Ease development by prepending to sys.path instead of appending
When sftpman-gtk is already installed as a package on the system,
and one tries to run `sftpman_gtk/launcher.py` during development,
the global files are going to be loaded, not the ones we want.

Appending to sys.path doesn't help, because the global packages path appears
first either way. It only works if there is no globally installed package.

Prepending fixes it for all cases.