Use scrolled window to prevent unlimited window height growth
When there are lots of sftp systems to show, the list would expand
and the whole window would grow in height indefinitely, not even
constrained by the size of the screen.

Using a scrolled window allows us to have a fixed size viewport
for the list and scrollbars when necessary.