Replace Mount/Unmount buttons with a Switch
The downside is that you can no longer "double click" in the same place
on the "Mount" button (which used to be replaced with an "Open" button)
to trigger mounting and opening.

You now need to mount via the Switch to the left and then click the Open
button in a different place (towards the middle of the row).